modified LedConfig0 argument type

+++ boards/stm32l476rg/src/tasks/led_driver.rs
@@ -28,7 +28,7 @@ pub async fn write_led() -> ! {
             };
         }
         match led_driver_sensor.set_led_colour(
-            LedDriverConfigAddresses::LedConfig0 as u8,
+            LedDriverConfigAddresses::LedConfig0,
             DRIVER_ADDRESSES[index],
             0xFF,
             0xFF,
